requirements.txt
supabase 
google-api-python-client
google-auth-httplib2 
google-auth-oauthlib 
groq 
python-dotenv 
python-docx